API Reference

This page provides a list of all documented functions and structs in Atomistic.jl. The list is not structured in any meaningful way, so a more clear summary of the package structure can be found in the targetted subsections of this documentation or in the source code itself.

Atomistic.DynamicAtom — Type

DynamicAtom{D,L<:Unitful.Length,V<:Unitful.Velocity} <: AbstractAtom

An atom representation based on the StaticAtom but with a velocity.

Type parameters

D: the dimension of the coordinate space

L: the type for the position components

V: the type for the velocity components

Atomistic.DynamicSystem — Type

DynamicSystem{D,AT<:AbstractAtom,L<:Unitful.Length,TT<:Unitful.Time} <: AbstractAtomicSystem{D}

A representation of a system of dynamic atoms which is similar to the FlexibleSystem but with a time field.

Type parameters

D: the dimension of the coordinate space

A: the type for the atoms that make up the system

L: the type for the bounding box components

T: the type for time field

Atomistic.MolecularDynamicsResult — Type

MolecularDynamicsResult

Abstract type to be extended by all concrete structs representing the result of a molecular dynamics simulation.

Atomistic.MolecularDynamicsSimulator — Type

MolecularDynamicsSimulator

Abstract type to be extended by all concrete molecular dynamics simulators.

Atomistic.get_system — Function

get_system(result::MolecularDynamicsResult, t::Integer = 0)::AbstractSystem

Extract the underlying system at a particular timestep from the simulation result.

An implementer of this API should implement a method of this function for their custom result type.

Atomistic.get_time_range — Method

get_time_range(result::MolecularDynamicsResult)::Vector{<:Real}

Extract the time range from the simulation result in atomic units.

An implementer of this API should implement a method of this function for their custom result type.

Atomistic.kinetic_energy — Function

kinetic_energy(result::MolecularDynamicsResult, t::Integer = 0)::Real

Extract the kinetic energy of the simulation at a particular timestep from the simulation result in atomic units.

An implementer of this API should implement a method of this function for their custom result type.

Atomistic.potential_energy — Function

potential_energy(result::MolecularDynamicsResult, t::Integer = 0)::Real

Extract the potential energy of the simulation at a particular timestep from the simulation result in atomic units.

An implementer of this API should implement a method of this function for their custom result type.

Atomistic.rdf — Function

rdf(result::MolecularDynamicsResult, sample_fraction::Float64 = 1.0)::Tuple{Vector{<:Real},Vector{<:Real}}

Calculate the radial distribution function from the simulation result.

To include only a trailing portion of the timesteps for reduced noise and faster computation, set sample_fraction to be less than 1.0. The result is a tuple of vectors which represent the radial distances (in atomic units) and the value of the rdf at each distance respectively.

Atomistic.reference_temperature — Method

reference_temperature(result::MolecularDynamicsResult)::Union{Real,Missing}

Extract the reference temperature of the simulation from the simulation result. If there is no thermostat with a reference temperature in this simulation, return missing.

An implementer of this API should implement a method of this function for their custom result type if it supports thermostats. If not implmented, the default implemenation just returns missing.

Atomistic.simulate — Method

simulate(system::AbstractSystem, simulator::MolecularDynamicsSimulator, potential::ArbitraryPotential)::MolecularDynamicsResult

Run a molecular dynamics simulation configured with a particular simulator and potential with any abstract system.

An implementer of this API should implement a method of this function for their custom simulator type. If the simulator has a fast path for some types of potential, those should be implemented with multiple dispatch.

Atomistic.temperature — Function

temperature(result::MolecularDynamicsResult, t::Integer = 0)::Real

Extract the temperature of the simulation at a particular timestep from the simulation result in atomic units.

An implementer of this API should implement a method of this function for their custom result type.

Atomistic.total_energy — Function

total_energy(result::MolecularDynamicsResult, t::Integer = 0)::Real

Extract the total energy of the simulation at a particular timestep from the simulation result in atomic units.

The default implementation simply sums the kinetic and potential energy at the timestep. An implementer of this API could implement a method of this function for their custom result type if it supports a more efficient way to calculate this quantity.