API Reference

This page provides a list of all documented types and functions and in Atomistic.jl.

If you are looking for more specifics on the Atomistic API, see the Implementing the Atomistic API page.

Atomistic.DynamicSystem — Type

DynamicSystem{S<:AbstractSystem{D},T<:Unitful.Time} <: AbstractSystem{D}

Wraps any AbstractSystem with a time field to be able to cache expensive results.

Type parameters

D: the dimension of the coordinate space

S: the type of the wrapped system

T: the type for time field

Atomistic.MolecularDynamicsResult — Type

MolecularDynamicsResult

Abstract type to be extended by all concrete structs representing the result of a molecular dynamics simulation.

Atomistic.MolecularDynamicsSimulator — Type

MolecularDynamicsSimulator

Abstract type to be extended by all concrete molecular dynamics simulators.

Atomistic.NBSResult — Type

NBSResult <: MolecularDynamicsResult

The result generated from running an NBSimulator.

Field descriptions

result::SimulationResult the standard simulation result from NBodySimulator

Atomistic.NBSimulator — Type

NBSimulator <: MolecularDynamicsSimulator

A wrapper around NBodySimulator to implement the Atomistic API.

Field descriptions

Δt::Real the time between timesteps, assumed to be in atomic units

steps::Integer the number of timesteps for the simulation

t₀::Real the starting time of the simulation, assumed to be in atomic units; defaults to 0

thermostat::Thermostat the thermostat for the simulation; many options are defined by NBodySimulator, but a user could also define a custom thermostat; defaults to the NullThermostat

simulator::OrdinaryDiffEqAlgorithm the algorithm to be used for the ODE; defaults to VelocityVerlet

potentials::Dict{Symbol,PotentialParameters} dictionary of potentials; shouldn't be manipulated directly by the user

Atomistic.UnimplementedError — Type

UnimplementedError

Exception thrown in default implementation of API to indicate that an implementator did not provide an implementation of a particular API function.

Atomistic.get_system — Function

get_system(result::MolecularDynamicsResult, t::Integer = 0)::AbstractSystem

Extract the underlying system at a particular timestep from the simulation result.

An implementer of this API should implement a method of this function for their custom result type.

Atomistic.get_time_range — Method

get_time_range(result::MolecularDynamicsResult)::AbstractVector{Unitful.Time}

Extract the time range from the simulation result in atomic units.

An implementer of this API should implement a method of this function for their custom result type.

Atomistic.kinetic_energy — Function

kinetic_energy(result::MolecularDynamicsResult, t::Integer = 0)::Unitful.Energy

Extract the kinetic energy of the simulation at a particular timestep from the simulation result in atomic units.

An implementer of this API should implement a method of this function for their custom result type.

Atomistic.plot_energy! — Function

plot_energy!(p::Plot, result::MolecularDynamicsResult, stride::Integer, first_plot::Bool = false)::Plot

Plot the kinetic, potential, and total energy of a MolecularDynamicsResult against time, sampling every stride points. Add the new lines to an existing plot and update the legend only if it is the first plot. If it is not the first plot, add a vertical line to differentiate the segments of the simulation in the plot.

Atomistic.plot_energy — Method

plot_energy(result::MolecularDynamicsResult, stride::Integer)::Plot

Plot the kinetic, potential, and total energy of a MolecularDynamicsResult against time, sampling every stride points.

Atomistic.plot_rdf — Function

plot_rdf(result::MolecularDynamicsResult, σ::Unitful.Length, sample_fraction::Float64 = 1.0)::Plot

Plot the radial distribution function of a MolecularDynamicsResult sampling only the trailing sample_fraction of the timesteps. Use σ (from Lennard Jones) as a normalization factor for the radius.

Atomistic.plot_rdf — Function

plot_rdf(result::MolecularDynamicsResult, σ::Real, sample_fraction::Float64 = 1.0)::Plot

Plot the radial distribution function of a MolecularDynamicsResult sampling only the trailing sample_fraction of the timesteps. Use σ (from Lennard Jones) as a normalization factor for the radius (assumed to be in atomic units).

Atomistic.plot_temperature! — Function

plot_temperature!(p::Plot, result::MolecularDynamicsResult, stride::Integer, first_plot::Bool = false)::Plot

Plot the temperature of a MolecularDynamicsResult against time, sampling every stride points. Adds the new line to an existing plot and update the legend only if it is the first plot. If it is not the first plot, adds a vertical line to differentiate the segments of the simulation in the plot.

Atomistic.plot_temperature — Method

plot_temperature(result::MolecularDynamicsResult, stride::Integer)::Plot

Plot the temperature of a MolecularDynamicsResult against time, sampling every stride points.

Atomistic.potential_energy — Function

potential_energy(result::MolecularDynamicsResult, t::Integer = 0)::Unitful.Energy

Extract the potential energy of the simulation at a particular timestep from the simulation result in atomic units.

An implementer of this API should implement a method of this function for their custom result type.

Atomistic.rdf — Function

rdf(result::MolecularDynamicsResult, sample_fraction::Float64 = 1.0)::Tuple{AbstractVector{Real},AbstractVector{Real}}

Calculate the radial distribution function from the simulation result.

To include only a trailing portion of the timesteps for reduced noise and faster computation, set sample_fraction to be less than 1.0. The result is a tuple of vectors which represent the interparticle radial distances (in bohr) and the density of each distance respectively.

Atomistic.reference_temperature — Method

reference_temperature(result::MolecularDynamicsResult)::Union{Unitful.Temperature,Missing}

Extract the reference temperature of the simulation from the simulation result. If there is no thermostat with a reference temperature in this simulation, return missing.

An implementer of this API should implement a method of this function for their custom result type if it supports thermostats. If not implmented, the default implemenation just returns missing.

Atomistic.simulate — Method

simulate(system::AbstractSystem, simulator::MolecularDynamicsSimulator, potential::ArbitraryPotential)::MolecularDynamicsResult

Run a molecular dynamics simulation configured with a particular simulator and potential with any abstract system.

An implementer of this API should implement a method of this function for their custom simulator type. If the simulator has a fast path for some types of potential, those should be implemented with multiple dispatch.

Atomistic.temperature — Function

temperature(result::MolecularDynamicsResult, t::Integer = 0)::Unitful.Temperature

Extract the temperature of the simulation at a particular timestep from the simulation result in atomic units.

An implementer of this API should implement a method of this function for their custom result type.

Atomistic.total_energy — Function

total_energy(result::MolecularDynamicsResult, t::Integer = 0)::Unitful.Energy

Extract the total energy of the simulation at a particular timestep from the simulation result in atomic units.

The default implementation simply sums the kinetic and potential energy at the timestep. An implementer of this API could implement a method of this function for their custom result type if it supports a more efficient way to calculate this quantity.

Atomistic.write_nbs_animation — Method

write_nbs_animation(result::NBSResult, filename::String)

Animate an NBSResult and store the result in a .gif file.